# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CUENTY API - Enterprise
  description: |
    API REST completa para el sistema CUENTY de gestión de suscripciones para el mercado mexicano.
    
    ## Características
    - Autenticación JWT + API Keys
    - Rate limiting: 100 requests/minuto
    - Webhooks para integraciones
    - SDKs disponibles: Node.js, Python, PHP
    
    ## Ambiente
    - Base URL Producción: https://api.cuenty.com/v1
    - Base URL Sandbox: https://sandbox.api.cuenty.com/v1
    
  version: 3.0.0
  contact:
    name: CUENTY API Support
    email: api@cuenty.com
    url: https://cuenty.com/docs
  license:
    name: Propietario
    url: https://cuenty.com/legal/api-terms
servers:
  - url: https://api.cuenty.com/v1
    description: Servidor de Producción
  - url: https://sandbox.api.cuenty.com/v1
    description: Servidor de Sandbox
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para desarrolladores externos
  
  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          example: 1
        serviceName:
          type: string
          example: Netflix Premium
        provider:
          type: string
          example: Netflix
        status:
          type: string
          enum: [active, pending, expired, cancelled]
          example: active
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        monthlyPrice:
          type: number
          format: decimal
          example: 299.00
        currency:
          type: string
          example: MXN
    
    # Payment
...
paths:
...
  /users/me/subscriptions:
    get:
      tags:
        - Users
      summary: Obtener suscripciones del usuario
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, expired, cancelled]
      responses:
        '200':
          description: Lista de suscripciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                  total:
                    type: integer
  
  # ==========================================================================
  # SUBSCRIPTIONS
  # ==========================================================================
  /subscriptions:
    get:
      tags:
        - Subscriptions
      summary: Listar todas las suscripciones disponibles
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: provider
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de suscripciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
  /subscriptions/{id}:
    get:
      tags:
        - Subscriptions
      summary: Obtener detalles de una suscripción
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles de la suscripción
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: Suscripción no encontrada
  /subscriptions/{id}/activate:
    post:
      tags:
        - Subscriptions
      summary: Activar una suscripción
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Suscripción activada exitosamente
        '400':
          description: Error al activar suscripción
  
  # ==========================================================================
  # PAYMENTS
  # ==========================================================================
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []